+++ src/main/java/com/ktb10/munggaebe/auth/service/KakaoService.java
@@ -6,8 +6,8 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.ktb10.munggaebe.auth.dto.AccessTokenResponse;
 import com.ktb10.munggaebe.auth.dto.RefreshTokenResponse;
 import com.ktb10.munggaebe.auth.exception.OAuthResponseJsonProcessingException;
-import com.ktb10.munggaebe.auth.feign.KakaoAuthClient;
-import com.ktb10.munggaebe.auth.feign.KakaoMemberApiClient;
+import com.ktb10.munggaebe.auth.client.KakaoAuthClient;
+import com.ktb10.munggaebe.auth.client.KakaoMemberApiClient;
 import com.ktb10.munggaebe.auth.jwt.JwtTokenProvider;
 import com.ktb10.munggaebe.auth.jwt.TokenManager;
 import com.ktb10.munggaebe.auth.service.dto.LoginDto;
